requirements.txt
flask
requests
python-dotenv
youtube-transcript-api
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
